Add function to apply change, updated change request structure
You can now apply changes by doing a put request to the 'recycle_locations#update'.
The request is authenticated from a token with the secret of the respective user.

The change request structure is as mentioned no longer in 'recycle_locations#update'.
It felt relevant to make a model for the changes alone and therefore it's in 'recycle_location_changes#create'